data processing

+++ scripts/process_common_voice.py
@@ -40,7 +40,7 @@ def main(_):
             try:
                 if not os.path.exists(wav_file):
                     tfm.build(source_file, wav_file)
-                    y, sr = soundfile.read(wav_file, dtype=np.int16)
+                    y, sr = soundfile.read(wav_file, dtype=np.float32)
                     # y, sr = librosa.load(wav_file, sr=None)
                     yt, index = librosa.effects.trim(y, top_db=10)
                     yt = y[max(index[0] - 40000, 0): min(index[1] + 40000, len(y))]
